+++ CHANGES.md
@@ -6,6 +6,7 @@
 ### Added
 
 * Added `--assume-yes` option to the `launch` command. Use `--assume-yes` to tell Flintrock to automatically destroy the cluster if there are problems during launch.
+* Automatically retry Hadoop download from flaky Apache mirrors.
 
 
 ## 0.1.0 - 2015-12-11